docs. 문서 파일명 변경 및 위치 이동
