+++ src/components/globe/Globe.tsx
@@ -5,6 +5,7 @@ import mapboxgl from 'mapbox-gl';
 import { INITIAL_CENTER, ZOOM } from './Globe.content';
 import { pickImageMarker } from './globe.util';
 import * as St from './style';
+import { globeCluster } from '../../utile/GlobeCluster';
 import { useLikeStore } from '../../zustand/useLikeStore';
 import { useLocationStore } from '../../zustand/useLocationStore';
 import { useMapLocationStore } from '../../zustand/useMapLocationStore';
@@ -12,7 +13,6 @@ import { useMarkerInvisible } from '../../zustand/useMarkerInvisible';
 import { usePostStore } from '../../zustand/usePostStore';
 import { useSearchStore } from '../../zustand/useSearchStore';
 import { useModal } from '../common/overlay/modal/Modal.hooks';
-import { globeCluster } from '../globeCluster/GlobeCluster';
 
 import type { Tables } from '../../types/supabase';
