+++ src/util/GlobeCluster.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 
-import { loadPinImage, pinImages, removeMapLayersAndSource } from './GlobeCluster.util';
-import Detail from '../detail/Detail';
+import clusterFive from '../assets/pin/clusterFive.png';
+import clusterTen from '../assets/pin/clusterTen.png';
+import clusterTwenty from '../assets/pin/clusterTwenty.png';
+import unclusterPin from '../assets/pin/unclusterPin.png';
+import Detail from '../components/detail/Detail';
 
-import type { Tables } from '../../types/supabase';
+import type { Tables } from '../types/supabase';
 
 interface Props {
   mapLocation: any;
@@ -100,4 +103,26 @@ export const globeCluster = ({ clusterData, mapLocation, postsData, mount, flyTo
     mapLocation.on('click', 'cluster-pin', handleClusterPointClick);
     mapLocation.on('click', 'unclustered-point', handleUnclusteredPointClick);
   }
+};
+
+export const pinImages = [
+  { name: 'unclusterPin', url: unclusterPin },
+  { name: 'clusterFive', url: clusterFive },
+  { name: 'clusterTen', url: clusterTen },
+  { name: 'clusterTwenty', url: clusterTwenty },
+];
+
+export const loadPinImage = (mapLocation: any, imageName: string, imageUrl: string) => {
+  mapLocation?.loadImage(imageUrl, (error: Error | undefined, image: any) => {
+    if (error) throw error;
+    if (!mapLocation?.hasImage(imageName)) {
+      mapLocation?.addImage(imageName, image);
+    }
+  });
+};
+
+export const removeMapLayersAndSource = (map: any) => {
+  map.removeLayer('unclustered-point');
+  map.removeLayer('cluster-pin');
+  map.removeSource('pinPoint');
 };
